Extract TestNGMultiversionTest trait
