+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/testng/TestNGFilteringIntegrationTest.groovy
@@ -24,84 +24,7 @@ import org.gradle.testing.fixture.TestNGCoverage
 import spock.lang.Issue
 
 @TargetCoverage({ TestNGCoverage.SUPPORTED_BY_JDK })
-class TestNGFilteringIntegrationTest extends AbstractTestFilteringIntegrationTest {
-
-    @Override
-    BuildScriptConfiguration getBuildScriptConfiguration() {
-        return new BuildScriptConfiguration() {
-            @Override
-            String getTestFrameworkDependencies(String sourceSet) {
-                return """
-                    testImplementation 'org.testng:testng:${version}'
-                """.stripIndent()
-            }
-
-            @Override
-            String getConfigureTestFramework() {
-                return "useTestNG()"
-            }
-
-            @Override
-            String getIncludeCategoryOrTagConfigurationElement() {
-                // TODO implement this if needed
-                throw new UnsupportedOperationException()
-            }
-
-            @Override
-            String getExcludeCategoryOrTagConfigurationElement() {
-                // TODO implement this if needed
-                throw new UnsupportedOperationException()
-            }
-        }
-    }
-
-    @Override
-    TestSourceConfiguration getTestSourceConfiguration() {
-        return new TestSourceConfiguration() {
-            @Override
-            String getTestFrameworkImports() {
-                return """
-                    import org.testng.annotations.*;
-               """.stripIndent()
-            }
-
-            @Override
-            String getBeforeClassAnnotation() {
-                return "@BeforeClass"
-            }
-
-            @Override
-            String getAfterClassAnnotation() {
-                return "@AfterClass"
-            }
-
-            @Override
-            String getBeforeTestAnnotation() {
-                return "@BeforeTest"
-            }
-
-            @Override
-            String getAfterTestAnnotation() {
-                return "@AfterTest"
-            }
-
-            @Override
-            String getIgnoreOrDisabledAnnotation() {
-                return "@Ignore"
-            }
-
-            @Override
-            String getRunOrExtendWithAnnotation(String runOrExtendWithClasses) {
-                // TODO implement this if needed
-                throw new UnsupportedOperationException()
-            }
-
-            @Override
-            String maybeParentheses(String methodName) {
-                return methodName
-            }
-        }
-    }
+class TestNGFilteringIntegrationTest extends AbstractTestFilteringIntegrationTest implements TestNGMultiVersionTest {
 
     void theUsualFiles() {
         buildFile << """
